Refactored most of the code TODO: 1. Refactor RowActions.js       2. handle Edit state change by updating fetched row on stopEditMode
