+++ temp/src/components/Table/Actions/DeleteRow.js
@@ -6,7 +6,6 @@ const  deleteRowOnClick=(()=>{
     setFetchedRows((rows)=>rows.filter(r=>{ return !( r.id===deleteRow.id)}))
 })
     useEffect(()=>{
-        
         deleteRowOnClick()
     // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [deleteRow.id, setFetchedRows])
