--- temp/src/components/Table/Actions/EditButtons.js
@@ -1,80 +1,0 @@
-import { Box, Button } from '@mui/material';
-import { GridCellModes } from '@mui/x-data-grid';
-import React from 'react'
-
-function EditButtons(props) {
-    const { selectedCellParams, cellMode, cellModesModel, setCellModesModel } =
-        props;
-
-    const handleSaveOrEdit = () => {
-        if (!selectedCellParams) {
-            return;
-        }
-        const { id, field } = selectedCellParams;
-        if (cellMode === "edit") {
-            setCellModesModel({
-                ...cellModesModel,
-                [id]: { ...cellModesModel[id], [field]: { mode: GridCellModes.View } },
-            });
-        } else {
-            setCellModesModel({
-                ...cellModesModel,
-                [id]: { ...cellModesModel[id], [field]: { mode: GridCellModes.Edit } },
-            });
-        }
-    };
-
-    const handleCancel = () => {
-        if (!selectedCellParams) {
-            return;
-        }
-        const { id, field } = selectedCellParams;
-        setCellModesModel({
-            ...cellModesModel,
-            [id]: {
-                ...cellModesModel[id],
-                [field]: { mode: GridCellModes.View, ignoreModifications: true },
-            },
-        });
-    };
-
-    const handleMouseDown = (event) => {
-        // Keep the focus in the cell
-        event.preventDefault();
-    };
-
-    return (
-        <Box
-            sx={{
-                borderBottom: 1,
-                borderColor: "divider",
-                p: 1,
-
-            }}
-            style={{ display: 'flex', justifyContent: 'space-between' }}
-        >
-            <Box >
-                <Button
-                    onClick={handleSaveOrEdit}
-                    onMouseDown={handleMouseDown}
-                    disabled={!selectedCellParams}
-                    variant="outlined"
-                >
-                    {cellMode === "edit" ? "Save" : "Edit"}
-                </Button>
-                <Button
-                    onClick={handleCancel}
-                    onMouseDown={handleMouseDown}
-                    disabled={cellMode === "view"}
-                    variant="outlined"
-                    sx={{ ml: 1 }}
-                >
-                    Cancel
-                </Button>
-            </Box>
-
-        </Box>
-    )
-}
-
-export default EditButtons
