+++ temp/src/components/Table/TableData/FetchRows.js
@@ -6,8 +6,6 @@ import {endpoint} from '../../../App'
 import { useSnackbar } from 'notistack'
 
 
-
-
 function FetchRows(props) {
 const { enqueueSnackbar} = useSnackbar()
 
@@ -17,21 +15,19 @@ const { enqueueSnackbar} = useSnackbar()
   }, []);
   const performAPICall = async () => {
     const URL = endpoint;
-    // console.log(URL);
     try {
       const response = await axios.get(URL);
       const data = response.data;
-      // console.log(data);
       props.handleRowChange(data);
       return data;
     } catch (e) {
       if (e.response) {
-        console.log(e.response.message);
+        // console.log(e.response.message);
         enqueueSnackbar(e.response.message, { variant: "error" });
       } else {
-        console.log(e,
-          "Something went wrong. Check that the backend is running, reachable and returns valid JSON."
-        );
+        // console.log(e,
+        //   "Something went wrong. Check that the backend is running, reachable and returns valid JSON."
+        // );
         enqueueSnackbar(
           "Something went wrong. Check that the backend is running, reachable and returns valid JSON.",
           { variant: "error" }
